requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.20.0
alembic==1.16.5
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.20
email-validator==2.3.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyJWT==2.12.1
requests==2.31.0
